Created GobiertoIndicators module
