+++ app/controllers/gobierto_indicators/application_controller.rb
@@ -1,0 +1,5 @@
+module GobiertoIndicators
+  class ApplicationController < ::ApplicationController
+    layout "gobierto_indicators/layouts/application"
+  end
+end
