+++ app/controllers/gobierto_indicators/indicators_controller.rb
@@ -1,0 +1,8 @@
+module GobiertoIndicators
+  class IndicatorsController < GobiertoIndicators::ApplicationController
+    include User::SessionHelper
+
+    def index
+    end
+  end
+end
