+++ app/views/gobierto_indicators/layouts/application.html.erb
@@ -1,0 +1,20 @@
+<% content_for :javascript_include do %>
+  <%= javascript_tag do %>
+    window.populateData = {
+      token: "<%= Rails.application.secrets.tbi_api_token %>",
+      municipalityId: "<%= @site.place.id %>",
+      year: 2015
+    }
+  <% end %>
+  <%= javascript_include_tag 'gobierto_indicators/application', 'data-turbolinks-track' => true %>
+<% end %>
+
+<% content_for :breadcrumb_items do %>
+  <strong>
+    <%= link_to t('.budgets'), gobierto_budgets_budgets_path %>
+  </strong>
+  <span>/</span>
+  <%= yield(:breadcrumb_current_item) %>
+<% end %>
+
+<%= render template: "layouts/application" %>
